Added new changes for Parent profile.Also done changes from long to integer. Anoop

+++ InspireHome/src/main/java/com/inspire/startup/config/AuditingConfig.java
@@ -1,6 +1,8 @@
 package com.inspire.startup.config;
 
 
+import java.util.Optional;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
@@ -11,22 +13,20 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 import com.inspire.startup.service.CustomUserDetails;
 
-import java.util.Optional;
-
 @Configuration
 @EnableJpaAuditing
 public class AuditingConfig {
 
     @Bean
-    public AuditorAware<Long> auditorProvider() {
+    public AuditorAware<Integer> auditorProvider() {
         return new SpringSecurityAuditAwareImpl();
     }
 }
 
-class SpringSecurityAuditAwareImpl implements AuditorAware<Long> {
+class SpringSecurityAuditAwareImpl implements AuditorAware<Integer> {
 
     @Override
-    public Optional<Long> getCurrentAuditor() {
+    public Optional<Integer> getCurrentAuditor() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null ||
